[BE-72] test: Foo 삭제 후 MappingDummy 클래스 CustomObjectMapperTest 내부에 정의
